Create README.md
Bond Lab is a suite of software for the analysis of fixed income securities.  Bond Lab offers functions for the analysis and valuation of Mortgage and Asset Backed Securities. Bond Lab applications focus on the valuation, 
structuring, and risk management of structured product investments like mortgage backed securities, REMICs (CMOs), Interest Only Strips, . Bond Lab was written to support the book Investing in Mortgage and Asset Backed Securities with R and Bond Lab.